refactor: move navbar files to component dir
